TemporalConnection update propagation (#102)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
- Handle updates to TemporalConnection, for TWD's that are associated
with one, so that these updates are propagated deployments.
- This was achieved by adding the hashed contents of a
TemporalConnection as pod annotations. Thus, if a TemporalConnection
were to get updated, it would result in new deployments with the updated
annotation. The new updated secret would also be mounted so that the
workers starting can be started with the latest secrets.

## Why?
- core functionality

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->
- #10 

4. How was this tested:
- Added unit tests and also tested this locally. The only thing I was
not able to do was add an integration test since I was not able to
configure our temporaltest server using mTLS.

Here's how I tested this locally:
- I ran a worker pod which was using the `temporal-cloud-mtls` secret to
connect to Temporal.
- I updated the namespace the worker was connected to by changing the
mTLS secret. This resulted in the controller and the worker pods
throwing connection errors (expected)
- I updated the local secret in the local k8s cluster (and also changed
it's name to `temporal-cloud-mtls-1` since we want our controller to
also refresh the client it's using)
- Noticed that a rolling deployment was conducted and everything
stabalized.

Proof of a rolling update:
<img width="1056" height="58" alt="Screenshot 2025-08-07 at 5 41 34 PM"
src="https://github.com/user-attachments/assets/40720118-0bb5-4f85-aa80-2c7c257b65f2"
/>



5. Any docs updates needed?
- None