Keep dots in Build IDs and format Worker Deployment names `<k8s-namespace>/<twd-name>` (#133)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
Keep dots in Build IDs and format Worker Deployment names
`<k8s-namespace>/<twd-name>`

## Why?
Dots are allowed in pod names and its nice to maintain the semantic
version appearance of an image.
Namespace first means the same alphabetization as `kubectl list
--all-namespaces`.
Namespace first is more expected for a slash delimiter that implies
hierarchy

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
Unit tests and functional tests

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->