Only Delete Deployments of NotRegistered versions if TemporalState is non-empty (#147)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed?
Only Delete Deployments of NotRegistered versions if TemporalState is
non-empty

## Why?
TemporalState can be empty on the initial rollout of a Worker
Deployment, if no versioned workers have ever polled. In that case we
need to continue the reconcile loop to create the worker pods which will
poll and create the Worker Deployment.

TemporalState can also, rarely, be empty if the server returns a
transient NotFound error, which can happen at any time in a Worker
Deployment's life. If TemporalState is empty when there are Deprecated
Versions around, it is important that we do not take an empty
TemporalState to mean that those Deprecated Versions are NotRegistered
and should be deleted, because that would cause the controller to delete
those Deployments despite the fact that they are actually draining. This
change prevents that.

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
Unit test

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->