Update demo scripts (#145)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
<!-- Describe what has changed in this PR -->
- Creates a helm chart for the demo worker
- Fixes the build for the demo worker
- Updates go dependencies in the demo submodule
- Configures kube context and temporal cloud namespace via
`skaffold.env` (gitignored)
- Adds CI check to ensure that demo worker builds successfully

## Why?
<!-- Tell your future self why have you made these changes -->
These changes make it possible to run the demo in a custom k8s context
(not minikube) and against a configurable Temporal cloud namespace or
self-hosted cluster _without_ modifying checked-in files.

Also, the demo build was broken. 😅 

## Checklist
<!--- add/delete as needed --->

1. Closes #142

2. How was this tested:

Manually ran each command in the demo readme

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->

Demo readme includes a few updates.