Fix: Initial deployment without current version does not get promoted  (#148)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
- Right now, if someone were to deploy the first version of a fresh new
deployment, the version keeps on being Inactive and does not get
promoted to Current immediately.
- Note: This only happens if there is a gated workflow specified on a
TWD. This bug does not take place when there isn't a gated workflow.

## Why?
- We have the following invariant: _Note that when there is no current
version (as in an initial versioned worker deployment), the progressive
steps are skipped and v1 becomes the current version immediately. All
new workflow executions will now start on v1._

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
- Added two new unit tests.
- Also tested this out locally.
- Integration tests.

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->