Refactor and add integration tests (#150)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
- Refactored the integration tests present and added some more as well.
- The main refactoring was done in such a way that each type of test was
being tested against one of the three rollout strategies we have defined
for our controller.

## Why?
- More test coverage and also so that we can sleep better at night
knowing our controller is bullet proof

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
- Tests

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->