Move helm/temporal-worker-controller/templates/crds to helm/temporal-worker-controller/crds (#153)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
Move helm/temporal-worker-controller/templates/crds to
helm/temporal-worker-controller/crds

## Why?
The specially-handled helm `crds/` directory is supposed to be next to
`templates/`, not inside of it.
https://helm.sh/docs/topics/charts/
https://helm.sh/docs/chart_best_practices/custom_resource_definitions/

Upgrading CRDs is dangerous, so placing the CRDs outside of `templates/`
allows helm to handle them safely, as described in their page on Custom
Resource Definitions.

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
<!--- Please describe how you tested your changes/how we can test them
-->

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->