Bug Fix: Ignore LastModifierIdentity if server deleted a version for garbage collection (#163)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
Bug Fix: Ignore LastModifierIdentity if server deleted a version for
garbage collection

Also, fix issue in describe version that would sometimes cause a
deprecated version to be Drained with nil DrainedSince time.

## Why?
The server sets the `LastModifierIdentity` to
`"try-delete-for-add-version"` when it deletes old versions to do
garbage collection upon addition of the a worker version beyond the
maximum. We are changing this on the server and will patch v1.29 of the
OSS server to handle this more gracefully in v1.29.2 of the server,
however, for users who cannot or do not want to use the new patch of the
server immediately, this will fix the problem for them on the controller
side so that they won't encounter this if they're running against
Temporal Server v1.29.1 or v1.29.0.

The issue of nil DrainedSince time caused flakiness in tests that
involve Deprecated Drained versions. That flakiness appeared a lot in
the new test I added to test the bug fix, so I also fixed that.


## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
New functional test

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->